Test on older versions of Rust
Also test both the MSVC and GNU ABI on Windows for the nightly and beta
channel. The build matrix already suffers from combinatorial explosion,
but this should be sufficient to catch regressions. On Windows, Rust
versions before 1.6 do not put rustc in the path, so the Appveyor script
does not work properly.