Work around breaking change in libc crate
The libc crate started using the #[deprecated] attribute, which is not
supported on Rust 1.8.0 stable, and results in a compile error. This
change was presumably introduced in commit 13ce9f8 [1] as part of pull
request 859 [2].

The solution is to put Cargo.lock under source control and freeze all
dependency versions, to be more resilient to external changes.

[1]: https://github.com/rust-lang/libc/commit/13ce9f8b86068773a247afa8e8fe9513861016e6
[2]: https://github.com/rust-lang/libc/pull/859